+++ src/app/common/utility/index.ts
@@ -5,3 +5,4 @@ export { default as parseAnchor } from './parseAnchor';
 export { default as isRecord } from './isRecord';
 export { default as isRecordContainingKey } from './isRecordContainingKey';
 export { default as isMobile } from './isMobile';
+export { default as isApplePayWindow } from './is-apple-pay-window';