+++ src/app/customer/CheckoutButtonList.tsx
@@ -1,20 +1,24 @@
 import { CustomerInitializeOptions, CustomerRequestOptions } from '@bigcommerce/checkout-sdk';
 import React, { memo, Fragment, FunctionComponent } from 'react';
 
+import { isApplePayWindow } from '../common/utility';
 import { TranslatedString } from '../locale';
 
 import { ApplePayButton } from './customWalletButton';
 import CheckoutButton from './CheckoutButton';
 
+const APPLE_PAY = 'applepay';
+
 // TODO: The API should tell UI which payment method offers its own checkout button
 export const SUPPORTED_METHODS: string[] = [
     'amazon',
     'amazonpay',
-    'applepay',
+    APPLE_PAY,
     'braintreevisacheckout',
     'chasepay',
     'masterpass',
     'googlepayadyenv2',
+    'googlepayadyenv3',
     'googlepayauthorizenet',
     'googlepaybraintree',
     'googlepaycheckoutcom',
@@ -40,8 +44,13 @@ const CheckoutButtonList: FunctionComponent<CheckoutButtonListProps> = ({
     methodIds,
     ...rest
 }) => {
-    const supportedMethodIds = (methodIds ?? [])
-        .filter(methodId => SUPPORTED_METHODS.indexOf(methodId) !== -1);
+    const supportedMethodIds = (methodIds ?? []).filter(methodId => {
+        if (methodId === APPLE_PAY && !isApplePayWindow(window)) {
+            return false;
+        }
+
+        return SUPPORTED_METHODS.indexOf(methodId) !== -1;
+    });
 
     if (supportedMethodIds.length === 0) {
         return null;