+++ src/app/payment/payment-methods.mock.ts
@@ -11,6 +11,9 @@ export function getPaymentMethod(): PaymentMethod {
             'AMEX',
             'MC',
         ],
+        initializationData: {
+            payPalCreditProductBrandName: {credit: ''},
+        },
         config: {
             displayName: 'Authorizenet',
             cardCode: true,
@@ -21,6 +24,21 @@ export function getPaymentMethod(): PaymentMethod {
             isVisaCheckoutEnabled: undefined,
             merchantId: undefined,
             testMode: false,
+        },
+        type: 'PAYMENT_TYPE_API',
+    };
+}
+
+export function getPaypalCreditPaymentMethod(): PaymentMethod {
+    return {
+        config: {},
+        supportedCards: [],
+        id: 'paypalcommercecredit',
+        gateway: undefined,
+        logoUrl: '',
+        method: 'paypal',
+        initializationData: {
+            payPalCreditProductBrandName: {credit: 'Pay in 3'},
         },
         type: 'PAYMENT_TYPE_API',
     };