+++ src/app/payment/paymentMethod/PPSDKPaymentMethod/initializationComponentMap.ts
@@ -1,9 +1,12 @@
-import { FunctionComponent } from 'react';
+import { ComponentType } from 'react';
+
+import HostedCreditCardPaymentMethod from '../HostedCreditCardPaymentMethod';
 
 import { NoUI } from './NoUI';
 
-type ComponentMap = Record<string, FunctionComponent>;
+type ComponentMap = Record<string, ComponentType<any>>;
 
 export const initializationComponentMap: ComponentMap = {
+    card_ui: HostedCreditCardPaymentMethod,
     none: NoUI,
 };