+++ src/app/payment/paymentMethod/getPaymentMethodDisplayName.spec.tsx
@@ -1,7 +1,7 @@
 import { createLanguageService, LanguageService } from '@bigcommerce/checkout-sdk';
 
 import { FALLBACK_TRANSLATIONS } from '../../locale/translations';
-import { getPaymentMethod } from '../payment-methods.mock';
+import { getPaymentMethod, getPaypalCreditPaymentMethod } from '../payment-methods.mock';
 
 import getPaymentMethodDisplayName from './getPaymentMethodDisplayName';
 import PaymentMethodId from './PaymentMethodId';
@@ -18,6 +18,13 @@ describe('getPaymentMethodDisplayName()', () => {
 
         expect(getPaymentMethodDisplayName(language)(method))
             .toEqual(method.config.displayName);
+    });
+
+    it('returns translated "Pay in 3" display name', () => {
+        const method = { ...getPaypalCreditPaymentMethod() };
+
+        expect(getPaymentMethodDisplayName(language)(method))
+            .toEqual(method.initializationData.payPalCreditProductBrandName.credit);
     });
 
     it('returns translated "Credit card" display name', () => {