+++ src/app/payment/paymentMethod/getPaymentMethodDisplayName.tsx
@@ -9,8 +9,16 @@ export default function getPaymentMethodDisplayName(
         const { displayName } = method.config;
 
         const isCreditCard = displayName?.toLowerCase() === 'credit card';
+        if (method.id === PaymentMethodId.PaypalCommerceCredit || method.id === PaymentMethodId.BraintreePaypalCredit) {
+            const { payPalCreditProductBrandName } = method.initializationData;
+            if (payPalCreditProductBrandName) {
+                return payPalCreditProductBrandName.credit || payPalCreditProductBrandName;
+            }
 
-        if (isCreditCard && method.id === PaymentMethodId.AdyenV2) {
+            return 'Pay Later';
+        }
+
+        if (isCreditCard && method.id === PaymentMethodId.AdyenV2 || method.id === PaymentMethodId.AdyenV3) {
             return language.translate('payment.adyen_credit_debit_card_text');
         }